In the video- These sample files works until video 9:04/22:46
Just download the folder mvc-master and copy the folder into your root htdocs and and then make sure Apache is started and running. The visit the link http://localhost/mvc-master/ and you should see the work in progress.
I encourage you to take it from  9:04/22:46 on the video. Watch and do it while you test it until it is done. Let me know if you need additional help.